User KEY TYPE Integer로 변경
User엔티티 seq 타입이 Integer임

+++ demo/src/main/java/com/example/demo/controller/web/UserController.java
@@ -36,7 +36,7 @@ public class UserController {
     }
 
     @GetMapping("{id}")
-    public String updateFrom(@PathVariable Long id, Model model) {
+    public String updateFrom(@PathVariable Integer id, Model model) {
         User user = userService.findById(id);
         model.addAttribute("user", user);
